+++ hugr-core/src/builder/dataflow.rs
@@ -16,6 +16,11 @@ use crate::types::{PolyFuncType, Signature, Type};
 use crate::{Direction, Hugr, IncomingPort, Node, OutgoingPort, Visibility, Wire, hugr::HugrMut};
 
 /// Builder for a [`ops::DFG`] node.
+///
+/// See the [crate-level documentation](https://docs.rs/hugr/latest/hugr/#example)
+/// for an example of using [`DFGBuilder`] to construct a boolean circuit or the
+/// [extension module documentation](crate::extension) for an example of using the builder
+/// with a quantum extension.
 #[derive(Debug, Clone, PartialEq)]
 pub struct DFGBuilder<T> {
     pub(crate) base: T,