+++ hugr-core/src/extension.rs
@@ -1,7 +1,111 @@
-//! Extensions
+//! Extension framework for user-defined operations and types.
 //!
-//! TODO: YAML declaration and parsing. This should be similar to a plugin
-//! system (outside the `types` module), which also parses nested [`OpDef`]s.
+// TODO: YAML declaration and parsing. This should be similar to a plugin
+// system (outside the `types` module), which also parses nested [`OpDef`]s.
+//!
+//!
+//! # Example
+//!
+//! By default HUGR does not include any quantum operations, but it is possible to define
+//! them in an extension and then use them in a HUGR.
+//! Here we show how to define a custom quantum extension, and how to use it to build a HUGR for a simple quantum circuit.
+//! ```
+//! use hugr::builder::{BuildError, DFGBuilder, Dataflow, DataflowHugr, inout_sig};
+//! // The prelude includes basic types like `bool`, `usize_t` and a `qubit` type.
+//! use hugr::extension::prelude::{bool_t, qb_t};
+//! use hugr::envelope::EnvelopeConfig;
+//! use hugr::hugr::Hugr;
+//!
+//! // By default, no gateset is defined. This module provides Hadamard and CX gates.
+//! mod mini_quantum_extension {
+//!     use hugr::{
+//!         extension::{
+//!             prelude::{bool_t, qb_t},
+//!             ExtensionId, Version,
+//!         },
+//!         ops::{ExtensionOp, OpName},
+//!         types::{FuncValueType, PolyFuncTypeRV},
+//!         Extension,
+//!     };
+//!
+//!     use std::sync::{Arc, LazyLock};
+//!
+//!     fn one_qb_func() -> PolyFuncTypeRV {
+//!         FuncValueType::new_endo(vec![qb_t()]).into()
+//!     }
+//!
+//!     fn two_qb_func() -> PolyFuncTypeRV {
+//!         FuncValueType::new_endo(vec![qb_t(), qb_t()]).into()
+//!     }
+//!     /// The extension identifier.
+//!     pub const EXTENSION_ID: ExtensionId = ExtensionId::new_unchecked("mini.quantum");
+//!     pub const VERSION: Version = Version::new(0, 1, 0);
+//!     fn extension() -> Arc<Extension> {
+//!         Extension::new_arc(EXTENSION_ID, VERSION, |ext, extension_ref| {
+//!             ext.add_op(OpName::new_inline("H"), "Hadamard".into(), one_qb_func(), extension_ref)
+//!                 .unwrap();
+//!
+//!             ext.add_op(OpName::new_inline("CX"), "CX".into(), two_qb_func(), extension_ref)
+//!                 .unwrap();
+//!
+//!             ext.add_op(
+//!                 OpName::new_inline("Measure"),
+//!                 "Measure a qubit, returning the qubit and the measurement result.".into(),
+//!                 FuncValueType::new(vec![qb_t()], vec![qb_t(), bool_t()]),
+//!                 extension_ref,
+//!             )
+//!             .unwrap();
+//!         })
+//!     }
+//!
+//!     /// Quantum extension definition.
+//!     pub static EXTENSION: LazyLock<Arc<Extension>> = LazyLock::new(extension);
+//!
+//!     fn get_gate(gate_name: impl Into<OpName>) -> ExtensionOp {
+//!         EXTENSION
+//!             .instantiate_extension_op(&gate_name.into(), [])
+//!             .unwrap()
+//!             .into()
+//!     }
+//!     pub fn h_gate() -> ExtensionOp {
+//!         get_gate("H")
+//!     }
+//!
+//!     pub fn cx_gate() -> ExtensionOp {
+//!         get_gate("CX")
+//!     }
+//!
+//!     pub fn measure() -> ExtensionOp {
+//!         get_gate("Measure")
+//!     }
+//! }
+//!
+//! use mini_quantum_extension::{cx_gate, h_gate, measure};
+//!
+//! //      ┌───┐
+//! // q_0: ┤ H ├──■─────
+//! //      ├───┤┌─┴─┐┌─┐
+//! // q_1: ┤ H ├┤ X ├┤M├
+//! //      └───┘└───┘└╥┘
+//! // c:              ╚═
+//! fn make_dfg_hugr() -> Result<Hugr, BuildError> {
+//!     let mut dfg_builder = DFGBuilder::new(inout_sig(
+//!         vec![qb_t(), qb_t()],
+//!         vec![qb_t(), qb_t(), bool_t()],
+//!     ))?;
+//!     let [wire0, wire1] = dfg_builder.input_wires_arr();
+//!     let h0 = dfg_builder.add_dataflow_op(h_gate(), vec![wire0])?;
+//!     let h1 = dfg_builder.add_dataflow_op(h_gate(), vec![wire1])?;
+//!     let cx = dfg_builder.add_dataflow_op(cx_gate(), h0.outputs().chain(h1.outputs()))?;
+//!     let measure = dfg_builder.add_dataflow_op(measure(), cx.outputs().last())?;
+//!     dfg_builder.finish_hugr_with_outputs(cx.outputs().take(1).chain(measure.outputs()))
+//! }
+//!
+//! let h: Hugr = make_dfg_hugr().unwrap();
+//! // Serialize the hugr to obtain a printable representation
+//! let serialized = h.store_str(EnvelopeConfig::text()).unwrap();
+//! println!("{}", serialized);
+//! ```
 
 use itertools::Itertools;
 use resolution::{ExtensionResolutionError, WeakExtensionRegistry};
@@ -860,7 +964,6 @@ pub mod test {
         ) -> Arc<Self> {
             Self::new_arc(name, Version::new(0, 0, 0), init)
         }
-
         /// Create a new extension for testing, with a 0 version.
         pub(crate) fn try_new_test_arc(
             name: ExtensionId,